hosttagprovider: Move to a package
